CryptoComCCXT Trading Bot: Buys Random Sell Options
A crypto.com/exchange Crypto trading bot that Buys random amount of coin purchase it handles exceptions with ccxt bad.symbol to except unlisted Pairs from market it handles exceptions of insufficient balance with BaseError when low on USD! You can Clone this and use it by running the main.py file in a python program! or run "python3 main.py" in terminal when with python3 installed! the files need to be all in the same folder and path in the terminal and it should run!
You may want to adjust the execute_trade.py file to your own RISK at getting market profit values for a strategy!
if your going to use this be aware i take no liabillity to any loss or gains its all for your own Risk taking!